fixed bug that was causing duration to be retured as int
git-svn-id: https://ffmpeg-php.svn.sourceforge.net/svnroot/ffmpeg-php/trunk/ffmpeg-php@268 f31480fa-141b-0410-8f85-e1b837e7f62b

+++ ffmpeg_movie.c
@@ -308,7 +308,7 @@ static float _php_get_duration(ff_movie_context *ffmovie_ctx)
 {
     float duration;
 
-    duration = ffmovie_ctx->fmt_ctx->duration / AV_TIME_BASE;
+    duration = (float)ffmovie_ctx->fmt_ctx->duration / AV_TIME_BASE;
 
     if (duration < 0) {
         duration = 0.0f;
